fix: update CrawlerThreadTest and DataServiceImplTest for API changes
- CrawlerThreadTest: Replace setActiveThreadCount with incrementAndGetActiveThreadCount
- CrawlerThreadTest: Fix assertEquals to use int instead of Integer
- DataServiceImplTest: Add explicit cast for getId() return value

+++ fess-crawler/src/test/java/org/codelibs/fess/crawler/service/impl/DataServiceImplTest.java
@@ -108,7 +108,7 @@ public class DataServiceImplTest extends PlainTestCase {
                         accessResult.setStatus(200);
                         accessResult.setUrl("http://example.com/" + threadIndex + "/" + i);
                         dataService.store(accessResult);
-                        ids.add(accessResult.getId());
+                        ids.add((Long) accessResult.getId());
                     }
                 } catch (final Throwable e) {
                     errors.add(e);
